network_security/components/data_validation.py
from network_security.entity.artifact_entity import DataIngestionArtifact, DataValidationArtifact
from network_security.entity.config_entity import DataValidationConfig
from network_security.exception.exception import NetworkSecurityException
from network_security.constants.training_pipeline import SCHEMA_FILE_PATH
from network_security.logging_modules.logger import logging
from network_security.utils.main_utils.utils import read_yaml_file, write_yaml
from scipy.stats import ks_2samp
import pandas as pd
import os
import sys


class DataValidation:
    def __init__(self, data_ingestion_articact: DataIngestionArtifact, data_validation_config: DataValidationConfig):
        try:
            self.data_ingestion_articact = data_ingestion_articact
            self.data_validation_config = data_validation_config
            self.schema_config = read_yaml_file(SCHEMA_FILE_PATH)
        except Exception as e:
            raise NetworkSecurityException(e, sys)
        

    @staticmethod
    def read_file(file_path) -> pd.DataFrame:
        try:
            return pd.read_csv(file_path)

        except Exception as e:
            raise NetworkSecurityException(e, sys)
        

    def validate_number_of_columns(self, dataframe: pd.DataFrame) -> bool:
        try:
            number_of_column = len(self.schema_config)
            logging.info(f"Required no. of columns = {number_of_column}")
            logging.info(f'DataFrame has columns : {len(dataframe.columns)}')

            if len(dataframe.columns) == number_of_column:
                return True
            return False


        except Exception as e:
            raise NetworkSecurityException(e, sys)
        
    def numerical_column_exists (self, dataframe: pd.DataFrame) -> bool:
        try:
            len_of_num_col_in_dataframe = len(dataframe.select_dtypes(include=['number']).columns)
            len_of_num_col_in_schema = len(self.schema_config.get('numerical_columns', []))

            logging.info(f"Numeric columns in schema: {len_of_num_col_in_schema}")
            logging.info(f"Numeric columns in dataframe: {len_of_num_col_in_dataframe}")

            if len_of_num_col_in_dataframe == len_of_num_col_in_schema:
                return True
            return False
            

        except Exception as e:
            raise NetworkSecurityException(e, sys)

    
    def detect_dataset_drift(self, base_df, current_df, threshold = 0.05) -> bool:
        try:
            status = True
            report = {}
            for column in base_df.columns:
                d1 = base_df[column]
                d2 = current_df[column]
                is_same_distance = ks_2samp(d1, d2)
                if threshold <= is_same_distance.pvalue:
                    is_found = False
                else:
                    is_found = True

                report.update({column:{
                    'p_value': float(is_same_distance.pvalue),
                    'drift_status' : is_found
                }})
            drift_report_file_path = self.data_validation_config.drift_report_file_path
            dir_path = os.path.dirname(drift_report_file_path)
            os.makedirs(dir_path, exist_ok= True)

            write_yaml(file_path= drift_report_file_path, content= report)
        
        except Exception as e:
            raise NetworkSecurityException(e, sys)



    def initiate_data_validation(self) -> DataValidationArtifact:
        try:
            train_file_path = self.data_ingestion_articact.train_file_path
            test_file_path = self.data_ingestion_articact.test_file_path


            #read the data from train and test
            train_dataframe = DataValidation.read_file(train_file_path)
            test_dataframe = DataValidation.read_file(test_file_path)

            # validate number of columns
            status = self.validate_number_of_columns(dataframe=train_dataframe)
            if not status:
                error_message = 'Train dataframe doesnot contain all columns \n'

            status = self.validate_number_of_columns(dataframe = test_dataframe)
            if not status:
                error_message =  'Test dataframe doesnot contain all columns'


            status = self.numerical_column_exists(dataframe= train_dataframe)
            if not status:
                error_message = 'Train Dataframe: Numerical column length problem'

            status = self.numerical_column_exists(dataframe= test_dataframe)
            if not status:
                error_message = 'Test Dataframe: Numerical column length problem'


            # checking for data drift
            status = self.detect_dataset_drift(base_df= train_dataframe, current_df= test_dataframe)
            dir_path = os.path.dirname(self.data_validation_config.valid_train_file_path)
            os.makedirs(dir_path, exist_ok= True)

            train_dataframe.to_csv(
                self.data_validation_config.valid_train_file_path, index= False, header= True
            )

            test_dataframe.to_csv(
                self.data_validation_config.valid_test_file_path, index= False, header= True
            )

            data_validation_artifact = DataValidationArtifact(
                validation_status = status,
                valid_train_file_path = self.data_ingestion_articact.train_file_path,
                valid_test_file_path = self.data_ingestion_articact.test_file_path,
                invalid_train_file_path = None,
                invalid_test_file_path = None,
                drift_report_file_path = self.data_validation_config.drift_report_file_path
            )
            return data_validation_artifact

        except Exception as e:
            raise NetworkSecurityException(e, sys)
        