network_security/utils/main_utils/utils.py
import yaml
from network_security.logging_modules.logger import logging
from network_security.exception.exception import NetworkSecurityException
import os, sys
import numpy as np
import dill
import pickle

def read_yaml_file(file_path: str) -> dict:
    try:
        with open(file_path, 'rb') as yaml_file:
            return yaml.safe_load(yaml_file)

    except Exception as e:
        raise NetworkSecurityException(e, sys)
    
def write_yaml(file_path: str, content: object, replace: bool = False) -> None:
    try:
        if replace:
            if os.path.exists(file_path):
                os.remove(file_path)
        os.makedirs(os.path.dirname(file_path), exist_ok= True)
        with open (file_path, 'w') as file:
            yaml.dump(content, file)


    except Exception as e:
        raise NetworkSecurityException(e, sys)