requirements.txt
numpy
pandas
python-dotenv
setuptools
pymongo
certifi
pymongo[srv]==3.12
dotenv
scikit_learn
pyaml
dill


#-e .